Backend/application_service/requirements.txt
django
djangorestframework
mongoengine
django-rest-framework-mongoengine
djangorestframework-simplejwt[crypto]
requests
python-decouple
pika


#  python -m pip install -r .\requirements.txt